api/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pandas>=2.0.0
python-multipart>=0.0.9
pyyaml>=6.0
anthropic>=0.28.0
python-dotenv>=1.0.0
